search-api: Fix search-aggregation regression
Fixing the linting to not do implicit widening of Int->Long made the
cast in `handleBucketResult` fail.

This patch fixes that and also introduces a test that would catch this
in case of regressions :^)

+++ search-api/src/main/scala/no/ndla/searchapi/model/domain/TermAggregation.scala
@@ -7,7 +7,7 @@
 
 package no.ndla.searchapi.model.domain
 
-case class Bucket(value: String, count: Long)
+case class Bucket(value: String, count: Int)
 case class TermAggregation(
     field: Seq[String],
     sumOtherDocCount: Int,
